Refactor ExchangeParams and add constants for improved organization and clarity in API request parameters
